Rename report package
