+++ EventPlanner/app/src/main/java/m3/eventplanner/activities/MainActivity.java
@@ -89,6 +89,7 @@ public class MainActivity extends AppCompatActivity {
         topLevelDestinations.add(R.id.calendarFragment);
         topLevelDestinations.add(R.id.reservationConfirmationFragment);
         topLevelDestinations.add(R.id.commentApprovalFragment);
+        topLevelDestinations.add(R.id.reportFragment);
 
         navController = Navigation.findNavController(this, R.id.fragment_nav_content_main);
